requirements_test.txt
pytest-django
model-mommy
freezegun
coverage
tox>=1.7.0
